@@ -1,9 +1,7 @@
 import React from 'react';
-import './App.css';
+
 import MainComponent from './Components/MainComponent';
 
-function App() {
-  return (<MainComponent/>);
-}
+const App = () => < MainComponent />;
 
 export default App;
